requirements.txt
streamlit>=1.28.0
openai>=1.0.0
SpeechRecognition>=3.10.0
pyaudio>=0.2.14
pyttsx3>=2.90
requests>=2.31.0
python-dotenv>=1.0.0
groq>=0.4.0
huggingface-hub>=0.20.0
anthropic>=0.18.0

